backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
aiosmtplib
email-validator
lunar-python
pytz
pytest
pytest-asyncio
httpx
pgvector
langgraph
langchain
langchain-openai
aiosqlite
fakeredis
python-docx
pypdf